+++ packages/python-packages/api-stub-generator/tests/docstring_parser_test.py
@@ -5,7 +5,7 @@
 # --------------------------------------------------------------------------
 
 from apistub.nodes import DocstringParser
-from apistub.nodes import ArgType
+
 
 docstring_standard_return_type = """
 Dummy docstring to verify standard return types and param types
@@ -17,7 +17,7 @@ Dummy docstring to verify standard return types and param types
 :rtype: Union[str, int]
 """
 
-docstring_Union_return_type2 = """
+docstring_union_return_type_followed_by_irrelevant_text = """
 Dummy docstring to verify standard return types and param types
 :rtype: Union(str, int)
 Dummy string at new line
@@ -44,7 +44,7 @@ docstring_param_type = """
 :type val: str
 """
 
-docstring_param_type1 = """
+docstring_param_type_split = """
 :param str name: Dummy name param
 :param val: Value type
 :type val: str
@@ -65,6 +65,20 @@ docstring_param_nested_union = """
 :param dummyarg: Optional group to check if user exists in.
 :type dummyarg: typing.Union[~azure.eventhub.EventDataBatch, List[~azure.eventhub.EventData]]
 """
+
+docstring_param_multi_line_type_followed_by_irrelevant_text = """
+:param dummyarg: Optional group to check if user exists in.
+:type dummyarg: typing.Union[~azure.eventhub.EventDataBatch,
+    List[~azure.eventhub.EventData]]
+.. admonition:: Example:
+.. literalinclude:: ../samples/sample_detect_language.py
+    :start-after: [START batch_detect_language]
+    :end-before: [END batch_detect_language]
+    :language: python
+    :dedent: 8
+    :caption: Detecting language in a batch of documents.
+"""
+
 
 docstring_multi_complex_type = """
         :param documents: The set of documents to process as part of this batch.
@@ -104,33 +118,69 @@ docstring_param_type_private = """
 :type client: ~azure.search.documents._search_index_document_batching_client_base.SearchIndexDocumentBatchingClientBase
 """
 
+class TestDocstringParser:
 
-class TestDocStringParser:
+    def _test_variable_type(self, docstring, expected):
+        parser = DocstringParser(docstring)
+        for varname, expect_val in expected.items():
+            assert expect_val == parser.type_for(varname)
 
     def _test_return_type(self, docstring, expected):
-        docstring_parser = DocstringParser(docstring)
-        assert expected == docstring_parser.find_return_type()
-
-    def _test_variable_type(self, docstring, varname, expected):
-        docstring_parser = DocstringParser(docstring)
-        assert expected == docstring_parser.find_type("(type|keywordtype|paramtype|vartype)", varname)
-
-    def _test_find_args(self, docstring, expected_args, is_keyword = False):
         parser = DocstringParser(docstring)
-        expected = {}
-        for arg in expected_args:
-            expected[arg.argname] = arg
-        for arg in parser.find_args('keyword' if is_keyword else 'param'):
-            assert arg.argname in expected and arg.argtype == expected[arg.argname].argtype
-            
+        assert expected == parser.ret_type
+
+    def test_docstring_param_type(self):
+        self._test_variable_type(docstring_param_type, {
+            "name": "str",
+            "val": "str"
+        })
+
+    def test_docstring_param_type_split(self):
+        self._test_variable_type(docstring_param_type_split, {
+            "name": "str",
+            "val": "str",
+            "pipeline": "~azure.core.pipeline",
+            "pipe_id": "Union[str, int]",
+            "data": "str or ~azure.dummy.datastream"
+        })
+
+    def test_docstring_param_typing_optional(self):
+        self._test_variable_type(docstring_param_typing_optional, {
+            "group": "typing.Optional[str]"
+        })
+
+    def test_docstring_param_nested_union(self):
+        self._test_variable_type(docstring_param_nested_union, {
+            "dummyarg": "typing.Union[~azure.eventhub.EventDataBatch, List[~azure.eventhub.EventData]]"
+        })
+
+    def test_docstring_param_multi_line_type_followed_by_irrelevant_text(self):
+        self._test_variable_type(docstring_param_multi_line_type_followed_by_irrelevant_text, {
+            "dummyarg": "typing.Union[~azure.eventhub.EventDataBatch, List[~azure.eventhub.EventData]]"
+        })
+
+    def test_docstring_multi_complex_type(self):
+        self._test_variable_type(docstring_multi_complex_type, {
+            "documents": "list[str] or list[~azure.ai.textanalytics.DetectLanguageInput] or list[dict[str, str]]",
+            "country_hint": "str",
+            "model_version": "str",
+            "show_stats": "bool"
+        })
+
+    def test_docstring_param_type_private(self):
+        self._test_variable_type(docstring_param_type_private, {
+            "name": "str",
+            "client": "~azure.search.documents._search_index_document_batching_client_base.SearchIndexDocumentBatchingClientBase"
+        })
+
     def test_return_builtin_return_type(self):
         self._test_return_type(docstring_standard_return_type, "str")
 
     def test_return_union_return_type(self):
         self._test_return_type(docstring_Union_return_type1, "Union[str, int]")
     
-    def test_return_union_return_type1(self):
-        self._test_return_type(docstring_Union_return_type2, "Union(str, int)")
+    def test_return_union_return_type_followed_by_irrelevant_text(self):
+        self._test_return_type(docstring_union_return_type_followed_by_irrelevant_text, "Union(str, int)")
 
     def test_return_union_lower_case_return_type(self):
         self._test_return_type(docstring_union_return_type3, "union[str, int]")
@@ -140,29 +190,3 @@ class TestDocStringParser:
 
     def test_dict_return_type(self):
         self._test_return_type(docstring_dict_ret_type, "dict[str, int]")
-
-    def test_param_type(self):
-        self._test_variable_type(docstring_param_type, "val", "str")
-
-    def test_param_type_private(self):
-        self._test_variable_type(docstring_param_type_private, "client", "~azure.search.documents._search_index_document_batching_client_base.SearchIndexDocumentBatchingClientBase")
-
-    def test_params(self):
-        args = [ArgType("name", "str"), ArgType("val", "str")]
-        self._test_find_args(docstring_param_type, args)
-    
-    def test_param_optional_type(self):
-        self._test_variable_type(docstring_param_type1, "pipe_id", "Union[str, int]")
-
-    def test_param_or_type(self):
-        self._test_variable_type(docstring_param_type1, "data", "str or ~azure.dummy.datastream")
-        self._test_variable_type(docstring_param_type1, "pipeline", None)
-
-    def test_type_typing_optional(self):
-        self._test_variable_type(docstring_param_typing_optional, "group", "typing.Optional[str]")
-
-    def test_nested_union_type(self):
-        self._test_variable_type(docstring_param_nested_union, "dummyarg", "typing.Union[~azure.eventhub.EventDataBatch, List[~azure.eventhub.EventData]]")
-
-    def test_multi_text_analytics_type(self):
-        self._test_variable_type(docstring_multi_complex_type, "documents", "list[str] or list[~azure.ai.textanalytics.DetectLanguageInput] or list[dict[str, str]]")