+++ packages/python-packages/api-stub-generator/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
 
+## Version 0.2.6 (Unreleased)
+Updated type parsing to properly handle cases in which type
+  info wraps onto a second line.
+Fixed issue where ivar type was not properly parsed from
+  the docstring.
+Fixed issue where typehint parsing was too strict and required
+  spaces.
+
 ## Version 0.2.5 (Unreleased)
 Fixed bug in which kwargs were duplicated in certain instances.
 Fix issue where non-async methods were erroneously marked async.